requirements.txt
# Stable Price Action Bot + MT5 + AI
# =================================

# Core Trading Dependencies
MetaTrader5==5.0.45
numpy<2.0
pandas==2.1.4
python-telegram-bot==20.7
python-dotenv==1.0.0
ta==0.10.2
matplotlib==3.8.2
seaborn==0.13.0
schedule==1.2.0
requests==2.31.0

# AI/ML Core Libraries
scikit-learn==1.3.2
scipy==1.9.3

# Advanced ML Libraries
xgboost==1.7.6
lightgbm==3.3.5

# Deep Learning / AI (CPU version to avoid DLL errors on Windows)
tensorflow-cpu==2.12.0
textblob==0.17.1
transformers==4.30.2

# Model Persistence & Configuration
joblib==1.3.2
pyyaml==6.0.1

# Additional AI Dependencies
numba==0.56.4
optuna==3.2.0